backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
openai
pandas
python-dotenv
pydantic
